feat: 🎸 材料の個数を改変できるようにした
個数がHelperから取得できない場合があるので

+++ src/main/java/dev/bluesheep/xeiexporter/api/XEIExporterApi.kt
@@ -30,7 +30,7 @@ abstract class XEIExporterApi {
     abstract fun createRecipeStackData(
         type: String,
         entry: String,
-        amount: Int = 1,
+        amount: Long = 1,
         chance: Float = 1f
     ): IRecipeStackData
 }
