+++ src/main/java/dev/bluesheep/xeiexporter/api/recipe/IRecipeStackData.kt
@@ -3,6 +3,6 @@ package dev.bluesheep.xeiexporter.api.recipe
 interface IRecipeStackData {
     val type: String
     val entry: String
-    val amount: Int
+    val amount: Long
     val chance: Float
 }
